DT-4 Refactoring integration test classes
