+++ src/main/java/br/com/ferracini/model/Speaker.java
@@ -13,7 +13,7 @@ import java.util.List;
 public class Speaker {
 
     @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @GeneratedValue(strategy = GenerationType.AUTO)
     @Column(name = "speaker_id")
     private Long speakerId;
 
@@ -39,6 +39,17 @@ public class Speaker {
     @ManyToMany(mappedBy = "speakers")
     @JsonIgnore
     private List<Session> sessions;
+
+    public Speaker() {
+    }
+
+    public Speaker(String firstName, String lastName, String title, String company, String speakerBio) {
+        this.firstName = firstName;
+        this.lastName = lastName;
+        this.title = title;
+        this.company = company;
+        this.speakerBio = speakerBio;
+    }
 
     public Long getSpeakerId() {
         return speakerId;
